chaged typeahead autocomplete for genre and language (#592)
For language terms, typing a language returns partially matched list of
items and form will select the first language in the list. This makes a
couple of adjustments. First the auto-complete (pressing tab) will not
assign a suggestion.  Second when focus leaves the input, if there are more
than one matches for what was typed, that value will remain what was typed and the
user will be forced to try again at the validation step (when they save).