Genre should maintain case (#589)
* Genre should maintain case

In coordination with #559, if you edit a Genre field with some invalid term
or alternative capitalization like "documentary" and save the form your
response will be accepted and not challenged. Investigation showed that
the validation does occur, but when there is alternative capitalization
it is accepted by the validation.  This change corrects the validation.

* fix and add tests